+++ src/components/Accomplishments/Accomplishments.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { Section, SectionDivider, SectionTitle } from '../../styles/GlobalComponents';
-import { Box, Boxes, BoxNum, BoxText } from './AcomplishmentsStyles';
+import { Box, Boxes, BoxNum, BoxText } from './AccomplishmentsStyles';
 
 const data = [
   { number: 15, text: 'Open Source Projects'},
@@ -10,7 +10,7 @@ const data = [
   { number: 12, text: 'Youtube Subscribers', }
 ];
 
-const Acomplishments = () => (
+const Accomplishments = () => (
   <Section>
     <SectionTitle>Personal Accomplishments</SectionTitle>
     <Boxes>
@@ -24,4 +24,4 @@ const Acomplishments = () => (
   </Section>
 );
 
-export default Acomplishments;
+export default Accomplishments;
