+++ src/components/BackgroundAnimation/BackgroundAnimation.js
